requirements.txt
typer>=0.12.3
rich>=13.7.1
PyGithub>=2.3.0
